Split IntegrationSpec in two
This commit makes the IntegrationSpec a trait that defines common
utilities used in most integration tests.

It splits the actual tests in two separate files that extend that trait.
The tests themselves aren't changed at all, except for the TCP ports they
use, and to remove nodes that have become useless in each of the specs.