fix override-features implementation
We were calling `nodeParams.features` from inside the channel, which is
problematic because we may have overriden those features for specific
peers. This is now fixed.

`IncomingPacket.decrypt` needs features to tell whether we are using
variable onion or not. What features should we use? Do we have an
unwritten rule that node-level features shouldn't be overriden? This
happens in the relayer and in the post-restart-htlc-cleaner.