Find the latest version of eclair-node jar
Until the build process is updated so that changing the major version number x to x+1 appends the `.0` to the new version number in the filename, Windows needs a different strategy to find the highest version number.  I chose to use the eclair-node*.jar file with the latest write time.  This is a fix for #1619 .