Correctly set gossip sync_complete
We are restoring the previous behavior of using the `sync_complete` field
to signal the end of a `channel_range_query` sync.

The first step is to correctly set that field, before we can read it and
interpret it to mark the end of sync.

See https://github.com/lightningnetwork/lightning-rfc/pull/826