try txs in hints before scanning the blockchain
This is a pretty sensitive part of the code. This implementation first
tries to get the txs provided in hints and verifies that one of them
indeed spend the watched tx (there can only be one because we watch a
single utxo). If (and only if) the spending tx checks out, then we
bypass the resource-intensive blockchain scanning.

Note: it still works if the hints are wrong.