Configurable anchor output max commitment feerate
It makes sense to allow node operators to configure the value they want
to use as a maximum threshold for anchor outputs commitment tx feerate.

This allows node operators to raise this value when mempools start getting
full in anticipation of a potential rise of the min-relay-fee.

This value can also be overridden for specific nodes.