flatten all versions, remove the legacy level
All versions are now at the same level, which means adding a new version
is really just adding a new package. This may remove a lot of headaches
for further migration.

The downside is that we can't restrict the visibility of `MigrationUtils`
(formerly `ChannelTypes`) to versions 0 and 1, but we would have had the
same issue as soon as we introduce version 3 anyway (because the current
version 2 codecs would have been moved to the legacy package).

```
wire
 |
 `-- internal
 |       |
 |       `-- channel
 |       |      |
 |       |      `-- MigrationTools
 |       |      |
 |       |      `-- version0
 |       |      |      |
 |       |      |      `-- ChannelCodecs0
 |       |      |
 |       |      `-- version1
 |       |      |      |
 |       |      |      `-- ChannelCodecs1
 |       |      |
 |       |      `-- version2
 |       |      |      |
 |       |      |      `-- ChannelCodecs2
 |       |      |
 |       |      `-- ChannelCodecs
 |       |
 |       `-- CommandCodecs
 |
 `-- others
```