set lock timeout at connection creation
From Hikari doc:
> 🔤connectionInitSql
This property sets a SQL statement that will be executed after every new
connection creation before adding it to the pool. If this SQL is not
valid or throws an exception, it will be treated as a connection failure
and the standard retry logic will be followed. Default: none