Rework the db version management (#1775)
* rework the db version management

The way our `getVersion`/`setVersion` was very unintuitive and led to
comments like the following in tests:
```scala
dbs.getVersion(statement, "network", 1) // this will set version to 1
```

The reason is that we treat unitialized databases and up-to-date
databases exactly the same way. That is why a `getVersion` will set the
version to the most up-to-date if none is found. It's also why we use
`CREATE IF NOT EXISTS` statements for tables and indices.

With this change, the `getVersion` now only _gets_ the version, or
returns `None` if no version exists.

Since we now differentiate between uninitialized and up-to-date db, we
don't need to make the create statements idempotent. This makes the code
more strict, and we will see errors if our versioning system has bugs.

Internal tables (used for versioning, locking, etc.) have been left
unchanged.

Co-authored-by: Bastien Teinturier <31281497+t-bast@users.noreply.github.com>