send command to itself to update relay fees at restore
This is an alternative to #1918 and #1920. It's very close to the
latter, except that we do check the conf only once in the
`WAIT_FOR_INIT_INTERNAL` state, as opposed to at each reconnection in
`SYNCING`.

We do not change the `channel_update` in `WAIT_FOR_INIT_INTERNAL`, which
allows us to set `previousChannelUpdate_opt=Some(normal.channelUpdate)`
in the transition and fix the duplicate bug in the audit db.

If there is a change in conf, there will be an additional
`LocalChannelUpdate` emitted, but only at reconnection, and following
the regular update flow, which should protect us against regressions.

We do handle `CMD_UPDATE_RELAY_FEES` in both `OFFLINE` and `SYNCING`,
because there may be a race between `CMD_UPDATE_RELAY_FEES` and
`ChannelRestablish` if the conf change at restore. And there was no good
reason to behave differently in those states anyway.