Remove  `messageFlags` from `ChannelUpdate` (#1941)
We make it a serialization detail, which it should be. The `derive`
method from `scodec` makes it very easy to do. We should probably always
use a dedicated class to handle flags, instead of using the `byte` codec
and binary operators.

This allows to remove the `require` in the `ChannelUpdate` definition,
which recently bit us in testing.

The only annoying thing is that we still need to expose a `messageFlags`
method in order to populate the `ChannelDisabled` error message.

We also typeify channel flags, as an alternative to passing around a `Byte`.