Add release notes in repository
It's useful to have the release notes directly inside the repository
instead of relying on Github: that lets for example users read them offline.

It also lets us fill the release notes every time we add relevant features
on `master` instead of having to do them all at once when we decide to make
a release.