Minor changes and refactoring (#1965)
* optionally record path-funding metrics

This is useful in feature branches.

Also rename `recordMetrics` to `recordPathFindingMetrics`.

* rename channel features attribute

It was a remainder of init features, which can be activated or not. But
in the context of the `ChannelFeatures` object, that naming was
confusing because all features are activated in that context.

* minor refactoring on channel_type

Moved some logic outside `Peer`.

* refactor `RouteParams`

`PathFindingConf` and `RouteParams` have almost the same set of params,
but some of them don't have the same structure (flat vs hierarchical).