new generic json serializer
Instead of providing a `MyClass => JValue` conversion method, we provide a `MyClass => MyClassJson` method, with the assumption that `MyClassJson` is serializable using the base serializers.

The rationale is that it's easier to define the structure with types rather than by building json objects.

It also has the advantage of removing recursion from custom serializers which sometimes generated weird stack overflows.