new generic json serializer
Instead of providing a `MyClass => JValue` conversion method, we
provide a `MyClass => MyClassJson` method, with the assumption that
`MyClassJson` is serializable using the base serializers.

The rationale is that it's easier to define the structure with types
rather than by building json objects.

This also means that the serialization of attributes of class C is out
of the scope when defining the serializer for class C. See for example
how `DirectedHtlcSerializer` doesn't need anymore to bring in
lower level serializers.

It also has the advantage of removing recursion from custom serializers
which sometimes generated weird stack overflows.