Update anchor outputs feerate tolerance (#1980)
Allow any feerate when using anchor outputs and we're fundee.
This will prevent unwanted channel closure.

This can be unsafe in a high fee environment if the commit tx is below
the propagation threshold. However, even when we discover it it's too late
anyway, so our only option is to wait for package relay to save the day.

Ensure feerate is always above propagation threshold when we're funder.
We lift the limit configured by the node operator when it is below the
network propagation threshold.