Fix API hanging on invalid remote params (#2008)
* reproduce bug causing API hang at open

In case of an error when validating channel parameters, we do not
return a message to the origin actor. That translates to API hanging
until timeout.

Took the opportunity to test return values in other cases too.

* return an error to origin actor for invalid params

* WaitForFundingCreatedInternal -> WaitForFundingInternal

* add tests to WaitForFundingInternalStateSpec

* add tests to WaitForFundingConfirmedStateSpec

* API nits

We probably don't need to print the stack trace for API errors, and the
open timeout of 10s was a bit short (it has to be << 30s though).