Add low-level route blinding features (#1962)
Cryptographic functions to blind and unblind a route and its associated
encrypted payloads.

Decrypt and decode the contents of an `encrypted_recipient_data` tlv field.
We could share the tlv namespace with onion tlvs, but it's cleaner to
separate them. They have a few common fields, but already diverge on
others, and will diverge even more in the future.