balance: take signed fulfills into account
For incoming htlc amounts need to be included in our balance if we know
the preimage, even if the htlc hasn't yet been formally settled.

We were already taking into accounts preimages in the `pending_commands`
database.

But, as soon as we have sent and signed an `update_fulfill_htlc`, we clean
up the `pending_commands` database. So we also need to look at current
sent changes.