Store blinding pubkey for introduction node (#2024)
It doesn't make sense to throw away this information, and it's useful in
some scenarios such as onion messages.

The ephemeral keys aren't part of the route, they're usually derived hop
by hop instead. We only need to keep the first one that must be somehow
sent to the introduction node.